pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trafficops"
version = "0.1.0"
description = "LLM-supervised adaptive traffic signal control with emergency prioritization"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "trafficops.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["trafficops", "trafficops.server"]
package-dir = { "trafficops" = ".", "trafficops.server" = "server" }
